requirements.txt
fastapi
uvicorn
python-dotenv
groq
sentence-transformers
scikit-learn
numpy
jinja2
python-multipart